+++ source/renderer/app/themes/daedalus/yellow.js
@@ -287,6 +287,7 @@ export const YELLOW_THEME_OUTPUT = {
     '--theme-loading-status-icons-unloaded-syncing-color': '#2d2d2d',
     '--theme-loading-status-icons-tooltip-color': '#2d2d2d',
     '--theme-loading-spinner-color': '#2d2d2d',
+    '--theme-loading-spinner-medium-color': '#2d2d2d',
   },
   manualUpdate: {
     '--theme-manual-update-overlay-background-color':
@@ -342,6 +343,37 @@ export const YELLOW_THEME_OUTPUT = {
       'rgba(45, 45, 45, 0.6)',
     '--theme-network-window-button-background-color-active':
       'rgba(45, 45, 45, 0.8)',
+  },
+  newsFeed: {
+    '--theme-news-feed-background-color': '#ffb923',
+    '--theme-news-feed-badge-background-color': '#ea4c5b',
+    '--theme-news-feed-badge-text-color': '#ffffff',
+    '--theme-news-feed-box-shadow-color': '-5px 0 20px 0 rgba(0, 0, 0, 0.25)',
+    '--theme-news-feed-header-background-color': '#eda70e',
+    '--theme-news-feed-header-title-color': '#2d2d2d',
+    '--theme-news-feed-icon-close-button-color': '#2d2d2d',
+    '--theme-news-feed-icon-close-hover-background-color': 'rgba(0, 0, 0, 0.1)',
+    '--theme-news-feed-icon-color': '#2d2d2d',
+    '--theme-news-feed-icon-color-connecting-screen': '#2d2d2d',
+    '--theme-news-feed-icon-color-syncing-screen': '#2d2d2d',
+    '--theme-news-feed-icon-dot-background-color': '#be0b0b',
+    '--theme-news-feed-icon-toggle-hover-background-color':
+      'rgba(0, 0, 0, 0.1)',
+    '--theme-news-feed-no-fetch-color': '#fafbfc',
+  },
+  newsItem: {
+    '--theme-news-item-action-button-background-color': 'rgba(0, 0, 0, 0.1)',
+    '--theme-news-item-action-button-background-color-hover': '#2d2d2d',
+    '--theme-news-item-action-button-border-color': '#2d2d2d',
+    '--theme-news-item-action-button-color': '#2d2d2d',
+    '--theme-news-item-action-button-color-hover': '#ffffff',
+    '--theme-news-item-alert-background-color': 'rgba(249, 128, 10, 0.5)',
+    '--theme-news-item-announcement-background-color':
+      'rgba(31, 193, 195, 0.3)',
+    '--theme-news-item-badge-color': '#ea4c5b',
+    '--theme-news-item-content-link-color': '#2d2d2d',
+    '--theme-news-item-info-background-color': 'rgba(0, 0, 0, 0.1)',
+    '--theme-news-item-title-color': '#2d2d2d',
   },
   nodeUpdate: {
     '--theme-node-update-background-color': '#f8f3ed',