+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 ### Features
 
+- Implemented "Newsfeed" feature ([PR 1570](https://github.com/input-output-hk/daedalus/pull/1570))
 - Implemented wallet recovery phrase verification ([PR 1565](https://github.com/input-output-hk/daedalus/pull/1565))
 - Removed select dropdown arrow ([PR 1550](https://github.com/input-output-hk/daedalus/pull/1550))
 - Implemented automated and manual update flows unification ([PR 1491](https://github.com/input-output-hk/daedalus/pull/1491))
@@ -44,7 +45,7 @@ Changelog
 
 ### Specifications
 
-- News feed ([PR 1569](https://github.com/input-output-hk/daedalus/pull/1569))
+- Newsfeed ([PR 1569](https://github.com/input-output-hk/daedalus/pull/1569))
 
 ## 0.14.0